refactor: move updateKeySigArrayForGrading into modifyKeySignature file
